requirements.txt
# Core data science packages
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
folium>=0.12.0

# Machine Learning
xgboost>=1.5.0
lightgbm>=3.2.0
joblib>=1.1.0

# Geospatial
geopandas>=0.10.0
shapely>=1.8.0
pyproj>=3.3.0

# Time series
statsmodels>=0.13.0

# Optimization
scipy>=1.7.0

# Utilities
tqdm>=4.62.0
pyyaml>=6.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=7.6.0

# Snowflake (if using locally)
snowflake-connector-python>=2.7.0
snowflake-sqlalchemy>=1.3.0
